[Finish] Finished

+++ src/pages/TodoPages.tsx
@@ -8,12 +8,14 @@ import {
   getTodos,
   searchTodos,
   toggleTodo,
+  updateTodo,
 } from "../services/todo.service";
 import { PiPlus } from "react-icons/pi";
 import { BiX } from "react-icons/bi";
 import { useAppDispatch, useAppSelector } from "../hooks/hooks";
 import { startLoading, stopLoading } from "../store/slices/uiSlice";
 import Loading from "../components/loading/Loading";
+import Swal from "sweetalert2";
 
 export default function TodoPages() {
   const loading = useAppSelector((s) => s.ui.globalLoading);
@@ -47,6 +49,13 @@ export default function TodoPages() {
   }, [debounce, statusFilter, categoryFilter]);
 
   const handleAddTodo = (name: string, category?: category) => {
+    Swal.fire({
+      title: "Todos Created!",
+      icon: "success",
+      timer: 800,
+      timerProgressBar: true,
+      showConfirmButton: false,
+    });
     addTodo(name, category);
     setTodos(getTodos());
   };
@@ -57,8 +66,34 @@ export default function TodoPages() {
   };
 
   const handleDelete = (id: number) => {
-    deleteTodo(id);
+    Swal.fire({
+      title: "Are you sure you want to delete this Item?",
+      text: "All the deleted items will be permanently deleted!",
+      icon: "question",
+      showConfirmButton: true,
+      showCancelButton: true,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Successfully deleted Item!",
+          icon: "success",
+          timer: 800,
+          showConfirmButton: false,
+        });
+        deleteTodo(id);
+        setTodos(getTodos());
+      }
+      if (result.dismiss || result.isDenied) return;
+    });
+  };
+
+  const handleEdit = (id: number, updates: Partial<Todo>) => {
+    updateTodo(id, updates);
     setTodos(getTodos());
+  };
+
+  const handleReorder = (newTodos: Todo[]) => {
+    setTodos(newTodos);
   };
 
   return (
@@ -68,7 +103,7 @@ export default function TodoPages() {
         <button
           className={`text-white transition flex ${
             !addForm ? "bg-green-400" : "bg-red-400"
-          } px-4 py-2 rounded-2xl mb-4 transition-colors duration-300`}
+          } px-4 py-2 rounded-2xl mb-4 transition-colors duration-300 cursor-pointer`}
           onClick={() => setAddForm(!addForm)}
         >
           {!addForm ? (
@@ -81,7 +116,7 @@ export default function TodoPages() {
         <button
           className={`text-white transition flex ${
             !searchTodo ? "bg-blue-400" : "bg-red-400"
-          } px-4 py-2 rounded-2xl mb-4 transition-colors duration-300`}
+          } px-4 py-2 rounded-2xl mb-4 transition-colors duration-300 cursor-pointer`}
           onClick={() => setSearchTodo(!searchTodo)}
         >
           {!searchTodo ? (
@@ -95,39 +130,41 @@ export default function TodoPages() {
       <div
         className={`overflow-hidden ${
           searchTodo ? "max-h-[200px]" : "max-h-0"
-        } transition-all ease-in-out duration-500 mb-4 space-x-2`}
+        } transition-all ease-in-out duration-500 mb-4 flex items-start justify-between`}
       >
         <input
           type="text"
           placeholder="Search todos..."
           value={search}
           onChange={(e) => setSearch(e.target.value)}
-          className="border p-2 mb-2"
+          className="border py-1 px-2 mb-2 rounded-2xl w-full"
         />
-        <select
-          value={statusFilter}
-          onChange={(e) => setStatusFilter(e.target.value as any)}
-          className="border p-2"
-        >
-          <option value="All">All</option>
-          <option value="Completed">Completed</option>
-          <option value="Pending">Pending</option>
-        </select>
+        <div className="flex flex-col items-end">
+          <select
+            value={statusFilter}
+            onChange={(e) => setStatusFilter(e.target.value as any)}
+            className="border py-1 px-2 mb-2 rounded-2xl"
+          >
+            <option value="All">All</option>
+            <option value="Completed">Completed</option>
+            <option value="Pending">Pending</option>
+          </select>
 
-        <select
-          value={categoryFilter}
-          onChange={(e) =>
-            setCategoryFilter(e.target.value as category | "All")
-          }
-          className="border p-2"
-        >
-          <option value="All">All Categories</option>
-          <option value="Career">Career</option>
-          <option value="Development">Development</option>
-          <option value="Learning">Learning</option>
-          <option value="Personal">Personal</option>
-          <option value="Work">Work</option>
-        </select>
+          <select
+            value={categoryFilter}
+            onChange={(e) =>
+              setCategoryFilter(e.target.value as category | "All")
+            }
+            className="border py-1 px-2 mb-2 rounded-2xl"
+          >
+            <option value="All">All Categories</option>
+            <option value="Career">Career</option>
+            <option value="Development">Development</option>
+            <option value="Learning">Learning</option>
+            <option value="Personal">Personal</option>
+            <option value="Work">Work</option>
+          </select>
+        </div>
       </div>
       <div
         className={`overflow-hidden ${
@@ -137,12 +174,14 @@ export default function TodoPages() {
         <TodoForm onAdd={handleAddTodo} />
       </div>
       {loading ? (
-        <Loading loadingText="Searching..."/>
+        <Loading loadingText="Searching..." />
       ) : (
         <TodoList
           todos={todos}
           handleDelete={handleDelete}
           handleToggle={handleToggle}
+          handleEdit={handleEdit}
+          handleReorder={handleReorder}
         />
       )}
     </div>
